Correct HVAC action and fix enum
According to the developer documentation at https://developers.home-assistant.io/docs/core/entity/climate, the HVAC action should more properly be the boiler state and not the zone state. This changes the HA integration to show the heating state only when the boiler is actually heating, and not when the EPH timer is set to heat, but the thermostat prohibits its activation. 

```
The HVAC action describes the current action. This is different from the mode, because if a device is set to heat, and the target temperature is already achieved, the device will not be actively heating anymore.
```

Also correct the enum use to use IntEnum